Review: Start the information hiding.

+++ Tree/BinarySearchTree/bin_search_tree.h
@@ -6,16 +6,16 @@
 #include <stdbool.h>
 
 /* Wrapper for BinSearchTree initialization. */
-#define BinSearchTree_init(p)       p = NULL; \
-                                    p = (BinSearchTree*)malloc(sizeof(BinSearchTree)); \
-                                    if (p != NULL) \
+#define BinSearchTree_init(p)       p = NULL;                                           \
+                                    p = (BinSearchTree*)malloc(sizeof(BinSearchTree));  \
+                                    if (p != NULL)                                      \
                                         BSTreeInit(p);     
 
 /* Wrapper for BinSearchTree deinitialization. */
-#define BinSearchTree_deinit(p)     if (p != NULL) { \
-                                        BSTreeDeinit(p); \
-                                        free(p); \
-                                        p = NULL; \
+#define BinSearchTree_deinit(p)     if (p != NULL) {                                    \
+                                        BSTreeDeinit(p);                                \
+                                        free(p);                                        \
+                                        p = NULL;                                       \
                                     }
 
 
@@ -32,14 +32,14 @@ typedef struct _BinSearchTree {
         
     TreeNode*     (*insert)     (struct _BinSearchTree*, void*);
     void          (*delete)     (struct _BinSearchTree*, TreeNode*);
-    unsigned long (*size)       (struct _BinSearchTree*);     
+    unsigned int  (*size)       (struct _BinSearchTree*);     
     bool          (*search)     (struct _BinSearchTree*, void*);
     TreeNode*     (*maximum)    (struct _BinSearchTree*);
     TreeNode*     (*minimum)    (struct _BinSearchTree*);
     TreeNode*     (*successor)  (struct _BinSearchTree*, TreeNode*);  
     TreeNode*     (*predecessor)(struct _BinSearchTree*, TreeNode*);
 
-    void          (*set_compare)(struct _BinSearchTree*, long (*)(const void*, const void*));
+    void          (*set_compare)(struct _BinSearchTree*, int (*)(const void*, const void*));
     void          (*set_destroy)(struct _BinSearchTree*, void (*) (void*));
 } BinSearchTree;
 
@@ -80,7 +80,7 @@ void BSTreeDelete(BinSearchTree *self, TreeNode *pNode);
  *
  * @return              The size;
  */
-unsigned long BSTreeSize(BinSearchTree *self);
+unsigned int BSTreeSize(BinSearchTree *self);
 
 
 /**
@@ -145,7 +145,7 @@ TreeNode* BSTreePredecessor(BinSearchTree *self, TreeNode *pCurNode);
  * @param self          The pointer to the BinSearchTree structure.
  * @param pFunc         The pointer to the customized function.
  */
-void BSTreeSetCompare(BinSearchTree *self, long (*pFunc)(const void*, const void*));
+void BSTreeSetCompare(BinSearchTree *self, int (*pFunc)(const void*, const void*));
 
 
 /**
